+++ changelog.md
@@ -70,3 +70,4 @@ Changed products service to work with server using Observable pattern
 
 17.03.2021
 Created orders service and make it work with server using Promise pattern
+Added TimingInterceptor for products reuests